+++ packages/block-editor/src/components/inserter/index.js
@@ -7,7 +7,7 @@ import { get } from 'lodash';
  */
 import { speak } from '@wordpress/a11y';
 import { __, _x, sprintf } from '@wordpress/i18n';
-import { Dropdown, IconButton } from '@wordpress/components';
+import { Dropdown, Button } from '@wordpress/components';
 import { Component } from '@wordpress/element';
 import { withDispatch, withSelect } from '@wordpress/data';
 import { compose, ifCondition } from '@wordpress/compose';
@@ -29,14 +29,10 @@ const defaultRenderToggle = ( { onToggle, disabled, isOpen, blockTitle, hasSingl
 		label = _x( 'Add block', 'Generic label for block inserter button' );
 	}
 	return (
-		<IconButton
+		<Button
 			icon="insert"
 			label={ label }
-			labelPosition="bottom"
-			onMouseDown={ ( event ) => {
-				event.preventDefault();
-				event.currentTarget.focus();
-			} }
+			tooltipPosition="bottom"
 			onClick={ onToggle }
 			className="block-editor-inserter__toggle"
 			aria-haspopup={ ! hasSingleBlockType ? 'true' : false }