+++ packages/block-editor/src/components/inserter/test/menu.js
@@ -8,91 +8,13 @@ import ReactDOM from 'react-dom';
 /**
  * Internal dependencies
  */
-import { InserterMenu, searchItems, normalizeTerm } from '../menu';
-
-const textItem = {
-	id: 'core/text-block',
-	name: 'core/text-block',
-	initialAttributes: {},
-	title: 'Text',
-	category: 'common',
-	isDisabled: false,
-	utility: 1,
-};
-
-const advancedTextItem = {
-	id: 'core/advanced-text-block',
-	name: 'core/advanced-text-block',
-	initialAttributes: {},
-	title: 'Advanced Text',
-	category: 'common',
-	isDisabled: false,
-	utility: 1,
-};
-
-const someOtherItem = {
-	id: 'core/some-other-block',
-	name: 'core/some-other-block',
-	initialAttributes: {},
-	title: 'Some Other Block',
-	category: 'common',
-	isDisabled: false,
-	utility: 1,
-};
-
-const moreItem = {
-	id: 'core/more-block',
-	name: 'core/more-block',
-	initialAttributes: {},
-	title: 'More',
-	category: 'layout',
-	isDisabled: true,
-	utility: 0,
-};
-
-const youtubeItem = {
-	id: 'core-embed/youtube',
-	name: 'core-embed/youtube',
-	initialAttributes: {},
-	title: 'YouTube',
-	category: 'embed',
-	keywords: [ 'google' ],
-	isDisabled: false,
-	utility: 0,
-};
-
-const textEmbedItem = {
-	id: 'core-embed/a-text-embed',
-	name: 'core-embed/a-text-embed',
-	initialAttributes: {},
-	title: 'A Text Embed',
-	category: 'embed',
-	isDisabled: false,
-	utility: 0,
-};
-
-const reusableItem = {
-	id: 'core/block/123',
-	name: 'core/block',
-	initialAttributes: { ref: 123 },
-	title: 'My reusable block',
-	category: 'reusable',
-	isDisabled: false,
-	utility: 0,
-};
-
-const items = [
-	textItem,
-	advancedTextItem,
-	someOtherItem,
-	moreItem,
-	youtubeItem,
-	textEmbedItem,
-	reusableItem,
-];
+import items, { categories, collections } from './fixtures';
+import { InserterMenu } from '../menu';
 
 const DEFAULT_PROPS = {
 	position: 'top center',
+	categories,
+	collections,
 	items,
 	debouncedSpeak: noop,
 	fetchReusableBlocks: noop,
@@ -163,7 +85,7 @@ describe( 'InserterMenu', () => {
 		const activeCategory = element.querySelector(
 			'.components-panel__body.is-opened > .components-panel__body-title'
 		);
-		expect( activeCategory.textContent ).toBe( 'Most Used' );
+		expect( activeCategory.textContent ).toBe( 'Most used' );
 	} );
 
 	it( 'should show nothing if there are no items', () => {
@@ -239,7 +161,7 @@ describe( 'InserterMenu', () => {
 
 	it( 'should show the common category blocks', () => {
 		const element = initializeAllClosedMenuStateAndReturnElement();
-		const commonBlocksTab = getTabButtonWithContent( element, 'Common Blocks' );
+		const commonBlocksTab = getTabButtonWithContent( element, 'Common blocks' );
 
 		TestUtils.Simulate.click( commonBlocksTab );
 
@@ -259,7 +181,7 @@ describe( 'InserterMenu', () => {
 
 	it( 'should disable items with `isDisabled`', () => {
 		const element = initializeMenuDefaultStateAndReturnElement();
-		const layoutTab = getTabButtonWithContent( element, 'Layout Elements' );
+		const layoutTab = getTabButtonWithContent( element, 'Layout elements' );
 
 		TestUtils.Simulate.click( layoutTab );
 
@@ -275,21 +197,21 @@ describe( 'InserterMenu', () => {
 		const element = initializeMenuDefaultStateAndReturnElement();
 		performSearchWithText( element, 'text' );
 
-		assertOpenedPanels( element, 2 );
+		assertOpenedPanels( element, 3 );
 
 		const matchingCategories = element.querySelectorAll(
 			'.components-panel__body-toggle'
 		);
 
-		expect( matchingCategories ).toHaveLength( 2 );
-		expect( matchingCategories[ 0 ].textContent ).toBe( 'Common Blocks' );
+		expect( matchingCategories ).toHaveLength( 3 );
+		expect( matchingCategories[ 0 ].textContent ).toBe( 'Common blocks' );
 		expect( matchingCategories[ 1 ].textContent ).toBe( 'Embeds' );
 
 		const visibleBlocks = element.querySelectorAll(
 			'.block-editor-block-types-list__item-title'
 		);
 
-		expect( visibleBlocks ).toHaveLength( 3 );
+		expect( visibleBlocks ).toHaveLength( 5 );
 		expect( visibleBlocks[ 0 ].textContent ).toBe( 'Text' );
 		expect( visibleBlocks[ 1 ].textContent ).toBe( 'Advanced Text' );
 		expect( visibleBlocks[ 2 ].textContent ).toBe( 'A Text Embed' );
@@ -301,71 +223,25 @@ describe( 'InserterMenu', () => {
 		const element = initializeMenuDefaultStateAndReturnElement();
 		performSearchWithText( element, ' text' );
 
-		assertOpenedPanels( element, 2 );
+		assertOpenedPanels( element, 3 );
 
 		const matchingCategories = element.querySelectorAll(
 			'.components-panel__body-toggle'
 		);
 
-		expect( matchingCategories ).toHaveLength( 2 );
-		expect( matchingCategories[ 0 ].textContent ).toBe( 'Common Blocks' );
+		expect( matchingCategories ).toHaveLength( 3 );
+		expect( matchingCategories[ 0 ].textContent ).toBe( 'Common blocks' );
 		expect( matchingCategories[ 1 ].textContent ).toBe( 'Embeds' );
 
 		const visibleBlocks = element.querySelectorAll(
 			'.block-editor-block-types-list__item-title'
 		);
 
-		expect( visibleBlocks ).toHaveLength( 3 );
+		expect( visibleBlocks ).toHaveLength( 5 );
 		expect( visibleBlocks[ 0 ].textContent ).toBe( 'Text' );
 		expect( visibleBlocks[ 1 ].textContent ).toBe( 'Advanced Text' );
 		expect( visibleBlocks[ 2 ].textContent ).toBe( 'A Text Embed' );
 
 		assertNoResultsMessageNotToBePresent( element );
-	} );
-} );
-
-describe( 'searchItems', () => {
-	it( 'should search items using the title ignoring case', () => {
-		expect( searchItems( items, 'TEXT' ) ).toEqual(
-			[ textItem, advancedTextItem, textEmbedItem ]
-		);
-	} );
-
-	it( 'should search items using the keywords', () => {
-		expect( searchItems( items, 'GOOGL' ) ).toEqual(
-			[ youtubeItem ]
-		);
-	} );
-
-	it( 'should search items using the categories', () => {
-		expect( searchItems( items, 'LAYOUT' ) ).toEqual(
-			[ moreItem ]
-		);
-	} );
-
-	it( 'should ignore a leading slash on a search term', () => {
-		expect( searchItems( items, '/GOOGL' ) ).toEqual(
-			[ youtubeItem ]
-		);
-	} );
-} );
-
-describe( 'normalizeTerm', () => {
-	it( 'should remove diacritics', () => {
-		expect( normalizeTerm( 'média' ) ).toEqual(
-			'media'
-		);
-	} );
-
-	it( 'should trim whitespace', () => {
-		expect( normalizeTerm( '  média  ' ) ).toEqual(
-			'media'
-		);
-	} );
-
-	it( 'should convert to lowercase', () => {
-		expect( normalizeTerm( '  Média  ' ) ).toEqual(
-			'media'
-		);
 	} );
 } );