+++ packages/block-library/src/paragraph/transforms.js
@@ -23,7 +23,7 @@ const transforms = {
 			} ),
 			transform( node ) {
 				const attributes = getBlockAttributes( name, node.outerHTML );
-				const { textAlign } = node.style;
+				const { textAlign } = node.style || {};
 
 				if (
 					textAlign === 'left' ||