+++ packages/block-library/src/video/edit.js
@@ -6,7 +6,6 @@ import {
 	BaseControl,
 	Button,
 	Disabled,
-	IconButton,
 	PanelBody,
 	ToolbarGroup,
 	withNotices,
@@ -186,8 +185,8 @@ class VideoEdit extends Component {
 			<>
 				<BlockControls>
 					<ToolbarGroup>
-						<IconButton
-							className="components-icon-button components-toolbar__control"
+						<Button
+							className="components-toolbar__control"
 							label={ __( 'Edit video' ) }
 							onClick={ switchToEditing }
 							icon="edit"