+++ packages/edit-post/src/components/header/writing-menu/index.js
@@ -22,21 +22,21 @@ function WritingMenu() {
 		>
 			<FeatureToggle
 				feature="fixedToolbar"
-				label={ __( 'Top Toolbar' ) }
+				label={ __( 'Top toolbar' ) }
 				info={ __( 'Access all block and document tools in a single place' ) }
 				messageActivated={ __( 'Top toolbar activated' ) }
 				messageDeactivated={ __( 'Top toolbar deactivated' ) }
 			/>
 			<FeatureToggle
 				feature="focusMode"
-				label={ __( 'Spotlight Mode' ) }
+				label={ __( 'Spotlight mode' ) }
 				info={ __( 'Focus on one block at a time' ) }
 				messageActivated={ __( 'Spotlight mode activated' ) }
 				messageDeactivated={ __( 'Spotlight mode deactivated' ) }
 			/>
 			<FeatureToggle
 				feature="fullscreenMode"
-				label={ __( 'Fullscreen Mode' ) }
+				label={ __( 'Fullscreen mode' ) }
 				info={ __( 'Work without distraction' ) }
 				messageActivated={ __( 'Fullscreen mode activated' ) }
 				messageDeactivated={ __( 'Fullscreen mode deactivated' ) }